+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 
 - 
 - 
+- Add `MissingCounter` metric ([#520](https://github.com/etna-team/etna/pull/520))
+- 
+- 
+- 
 - 
 - 
 -